requirements.txt
flake8
pyyaml
tox
